pulled out all options menus
