+++ src/frontend.cpp
@@ -81,6 +81,7 @@
 
 #include "uiscreens/titlemenu.h"
 #include "uiscreens/common.h"
+#include "uiscreens/options/graphicsoptions.h"
 
 // ////////////////////////////////////////////////////////////////////////////
 // Global variables
@@ -90,7 +91,6 @@ char			aLevelName[MAX_LEVEL_NAME_SIZE + 1];	//256];			// vital! the wrf file to 
 bool			bLimiterLoaded = false;
 
 #define TUTORIAL_LEVEL "TUTORIAL3"
-#define TRANSLATION_URL "https://translate.wz2100.net"
 
 // ////////////////////////////////////////////////////////////////////////////
 // Tutorial Menu
@@ -476,1683 +476,6 @@ bool runMultiPlayerMenu()
 
 	return true;
 }
-
-
-// ////////////////////////////////////////////////////////////////////////////
-// Options Menu
-void startOptionsMenu()
-{
-	sliderEnableDrag(true);
-
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("OPTIONS"));
-	addTextButton(FRONTEND_GAMEOPTIONS,	FRONTEND_POS2X, FRONTEND_POS2Y, _("Game Options"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_GRAPHICSOPTIONS, FRONTEND_POS3X, FRONTEND_POS3Y, _("Graphics Options"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_VIDEOOPTIONS, FRONTEND_POS4X, FRONTEND_POS4Y, _("Video Options"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_AUDIO_AND_ZOOMOPTIONS, FRONTEND_POS5X, FRONTEND_POS5Y, _("Audio / Zoom Options"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_MOUSEOPTIONS, FRONTEND_POS6X, FRONTEND_POS6Y, _("Mouse Options"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_KEYMAP,		FRONTEND_POS7X, FRONTEND_POS7Y, _("Key Mappings"), WBUT_TXTCENTRE);
-	addTextButton(FRONTEND_MUSICMANAGER, FRONTEND_POS8X, FRONTEND_POS8Y, _("Music Manager"), WBUT_TXTCENTRE);
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-	addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, _("Open Configuration Directory"), 0);
-}
-
-bool runOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_GAMEOPTIONS:
-		changeTitleMode(GAME);
-		break;
-	case FRONTEND_GRAPHICSOPTIONS:
-		changeTitleMode(GRAPHICS_OPTIONS);
-		break;
-	case FRONTEND_AUDIO_AND_ZOOMOPTIONS:
-		changeTitleMode(AUDIO_AND_ZOOM_OPTIONS);
-		break;
-	case FRONTEND_VIDEOOPTIONS:
-		changeTitleMode(VIDEO_OPTIONS);
-		break;
-	case FRONTEND_MOUSEOPTIONS:
-		changeTitleMode(MOUSE_OPTIONS);
-		break;
-	case FRONTEND_KEYMAP:
-		changeTitleMode(KEYMAP);
-		break;
-	case FRONTEND_MUSICMANAGER:
-		changeTitleMode(MUSIC_MANAGER);
-		break;
-	case FRONTEND_QUIT:
-		changeTitleMode(TITLE);
-		break;
-	case FRONTEND_HYPERLINK:
-		if (!openFolderInDefaultFileManager(PHYSFS_getWriteDir()))
-		{
-			// Failed to open write dir in default filesystem browser
-			std::string configErrorMessage = _("Failed to open configuration directory in system default file browser.");
-			configErrorMessage += "\n\n";
-			configErrorMessage += _("Configuration directory is reported as:");
-			configErrorMessage += "\n";
-			configErrorMessage += PHYSFS_getWriteDir();
-			configErrorMessage += "\n\n";
-			configErrorMessage += _("If running inside a container / isolated environment, this may differ from the actual path on disk.");
-			configErrorMessage += "\n";
-			configErrorMessage += _("Please see the documentation for more information on how to locate it manually.");
-			wzDisplayDialog(Dialog_Warning, _("Failed to open configuration directory"), configErrorMessage.c_str());
-		}
-		break;
-	default:
-		break;
-	}
-
-	// If close button pressed then return from this menu.
-	if (CancelPressed())
-	{
-		changeTitleMode(TITLE);
-	}
-
-	widgDisplayScreen(psWScreen);						// show the widgets currently running
-
-	return true;
-}
-
-char const *graphicsOptionsFmvmodeString()
-{
-	switch (war_GetFMVmode())
-	{
-	case FMV_1X: return _("1×");
-	case FMV_2X: return _("2×");
-	case FMV_FULLSCREEN: return _("Fullscreen");
-	default: return _("Unsupported");
-	}
-}
-
-char const *graphicsOptionsScanlinesString()
-{
-	switch (war_getScanlineMode())
-	{
-	case SCANLINES_OFF: return _("Off");
-	case SCANLINES_50: return _("50%");
-	case SCANLINES_BLACK: return _("Black");
-	default: return _("Unsupported");
-	}
-}
-
-char const *graphicsOptionsScreenShakeString()
-{
-	return getShakeStatus() ? _("On") : _("Off");
-}
-
-char const *graphicsOptionsSubtitlesString()
-{
-	return seq_GetSubtitles() ? _("On") : _("Off");
-}
-
-char const *graphicsOptionsShadowsString()
-{
-	return getDrawShadows() ? _("On") : _("Off");
-}
-
-char const *graphicsOptionsFogString()
-{
-	return pie_GetFogEnabled() ? _("On") : _("Off");
-}
-
-char const *graphicsOptionsRadarString()
-{
-	return rotateRadar ? _("Rotating") : _("Fixed");
-}
-
-char const *graphicsOptionsRadarJumpString()
-{
-	return war_GetRadarJump() ? _("Instant") : _("Tracked");
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Graphics Options
-void startGraphicsOptionsMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BOTFORM);
-
-	auto grid = std::make_shared<GridLayout>();
-	grid_allocation::slot row(0);
-
-	////////////
-	//FMV mode.
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_FMVMODE, _("Video Playback"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_FMVMODE_R, graphicsOptionsFmvmodeString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Scanlines
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_SCANLINES, _("Scanlines"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_SCANLINES_R, graphicsOptionsScanlinesString(), WBUT_SECONDARY)));
-	row.start++;
-
-	////////////
-	//shadows
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_SHADOWS, _("Shadows"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_SHADOWS_R, graphicsOptionsShadowsString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// fog
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_FOG, _("Fog"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_FOG_R, graphicsOptionsFogString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Radar
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_RADAR, _("Radar"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_RADAR_R, graphicsOptionsRadarString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// RadarJump
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_RADAR_JUMP, _("Radar Jump"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_RADAR_JUMP_R, graphicsOptionsRadarJumpString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// screenshake
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_SSHAKE, _("Screen Shake"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_SSHAKE_R, graphicsOptionsScreenShakeString(), WBUT_SECONDARY)));
-	row.start++;
-
-	grid->setGeometry(0, 0, FRONTEND_BUTWIDTH, grid->idealHeight());
-
-	auto scrollableList = ScrollableListWidget::make();
-	scrollableList->setGeometry(0, FRONTEND_POS2Y, FRONTEND_BOTFORMW - 1, FRONTEND_BOTFORMH - FRONTEND_POS2Y - 1);
-	scrollableList->addItem(grid);
-	parent->attach(scrollableList);
-
-	// Add some text down the side of the form
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("GRAPHICS OPTIONS"));
-
-	////////////
-	// quit.
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-}
-
-void seqFMVmode()
-{
-	war_SetFMVmode(seqCycle(war_GetFMVmode(), FMV_FULLSCREEN, 1, FMV_MODE(FMV_MAX - 1)));
-}
-
-void seqScanlineMode()
-{
-	war_setScanlineMode(seqCycle(war_getScanlineMode(), SCANLINES_OFF, 1, SCANLINES_BLACK));
-}
-
-bool runGraphicsOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_QUIT:
-		changeTitleMode(OPTIONS);
-		break;
-
-	case FRONTEND_SHADOWS:
-	case FRONTEND_SHADOWS_R:
-		setDrawShadows(!getDrawShadows());
-		widgSetString(psWScreen, FRONTEND_SHADOWS_R, graphicsOptionsShadowsString());
-		break;
-
-	case FRONTEND_FOG:
-	case FRONTEND_FOG_R:
-		if (pie_GetFogEnabled())
-		{
-			pie_SetFogStatus(false);
-			pie_EnableFog(false);
-		}
-		else
-		{
-			pie_EnableFog(true);
-		}
-		widgSetString(psWScreen, FRONTEND_FOG_R, graphicsOptionsFogString());
-		break;
-
-	case FRONTEND_FMVMODE:
-	case FRONTEND_FMVMODE_R:
-		seqFMVmode();
-		widgSetString(psWScreen, FRONTEND_FMVMODE_R, graphicsOptionsFmvmodeString());
-		break;
-
-	case FRONTEND_SCANLINES:
-	case FRONTEND_SCANLINES_R:
-		seqScanlineMode();
-		widgSetString(psWScreen, FRONTEND_SCANLINES_R, graphicsOptionsScanlinesString());
-		break;
-
-	case FRONTEND_RADAR:
-	case FRONTEND_RADAR_R:
-		rotateRadar = !rotateRadar;
-		widgSetString(psWScreen, FRONTEND_RADAR_R, graphicsOptionsRadarString());
-		break;
-
-	case FRONTEND_RADAR_JUMP:
-	case FRONTEND_RADAR_JUMP_R:
-		war_SetRadarJump(!war_GetRadarJump());
-		widgSetString(psWScreen, FRONTEND_RADAR_JUMP_R, graphicsOptionsRadarJumpString());
-		break;
-
-	case FRONTEND_SSHAKE:
-	case FRONTEND_SSHAKE_R:
-		setShakeStatus(!getShakeStatus());
-		widgSetString(psWScreen, FRONTEND_SSHAKE_R, graphicsOptionsScreenShakeString());
-		break;
-
-	default:
-		break;
-	}
-
-
-	// If close button pressed then return from this menu.
-	if (CancelPressed())
-	{
-		changeTitleMode(OPTIONS);
-	}
-
-	widgDisplayScreen(psWScreen);						// show the widgets currently running
-
-	return true;
-}
-
-static std::string audioAndZoomOptionsSoundHRTFMode()
-{
-	// retrieve whether the current "setting" is Auto or not
-	// as sound_GetHRTFMode() returns the current actual status (and never auto, even if the request was "auto")
-	bool isAutoSetting = war_GetHRTFMode() == HRTFMode::Auto;
-
-	std::string currentMode;
-	auto mode = sound_GetHRTFMode();
-	switch (mode)
-	{
-	case HRTFMode::Unsupported:
-		currentMode = _("Unsupported");
-		return currentMode;
-	case HRTFMode::Disabled:
-		currentMode = _("Disabled");
-		break;
-	case HRTFMode::Enabled:
-		currentMode = _("Enabled");
-		break;
-	case HRTFMode::Auto:
-		// should not happen, but if it does, just return
-		currentMode = _("Auto");
-		return currentMode;
-	}
-
-	if (isAutoSetting)
-	{
-		return std::string(_("Auto")) + " (" + currentMode + ")";
-	}
-	return currentMode;
-}
-
-static std::string audioAndZoomOptionsMapZoomString()
-{
-	char mapZoom[20];
-	ssprintf(mapZoom, "%d", war_GetMapZoom());
-	return mapZoom;
-}
-
-static std::string audioAndZoomOptionsMapZoomRateString()
-{
-	char mapZoomRate[20];
-	ssprintf(mapZoomRate, "%d", war_GetMapZoomRate());
-	return mapZoomRate;
-}
-
-static std::string audioAndZoomOptionsRadarZoomString()
-{
-	char radarZoom[20];
-	ssprintf(radarZoom, "%d", war_GetRadarZoom());
-	return radarZoom;
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Audio and Zoom Options Menu
-void startAudioAndZoomOptionsMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BOTFORM);
-
-	auto addSliderWrap = [](std::shared_ptr<WIDGET> widget) {
-		Alignment sliderAlignment(Alignment::Vertical::Center, Alignment::Horizontal::Left);
-		return sliderAlignment.wrap(addMargin(widget));
-	};
-
-	auto grid = std::make_shared<GridLayout>();
-	grid_allocation::slot row(0);
-
-	// 2d audio
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_FX, _("Voice Volume"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addSliderWrap(makeFESlider(FRONTEND_FX_SL, FRONTEND_BOTFORM, AUDIO_VOL_MAX, static_cast<UDWORD>(sound_GetUIVolume() * 100.0f))));
-	row.start++;
-
-	// 3d audio
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_3D_FX, _("FX Volume"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addSliderWrap(makeFESlider(FRONTEND_3D_FX_SL, FRONTEND_BOTFORM, AUDIO_VOL_MAX, static_cast<UDWORD>(sound_GetEffectsVolume() * 100.0f))));
-	row.start++;
-
-	// cd audio
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MUSIC, _("Music Volume"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addSliderWrap(makeFESlider(FRONTEND_MUSIC_SL, FRONTEND_BOTFORM, AUDIO_VOL_MAX, static_cast<UDWORD>(sound_GetMusicVolume() * 100.0f))));
-	row.start++;
-
-	////////////
-	//subtitle mode.
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_SUBTITLES, _("Subtitles"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_SUBTITLES_R, graphicsOptionsSubtitlesString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// HRTF
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_SOUND_HRTF, _("HRTF"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_SOUND_HRTF_R, audioAndZoomOptionsSoundHRTFMode(), WBUT_SECONDARY)));
-	row.start++;
-	if (sound_GetHRTFMode() == HRTFMode::Unsupported)
-	{
-		widgSetButtonState(psWScreen, FRONTEND_SOUND_HRTF, WBUT_DISABLE);
-		widgSetTip(psWScreen, FRONTEND_SOUND_HRTF, _("HRTF is not supported on your device / system / OpenAL library"));
-		widgSetButtonState(psWScreen, FRONTEND_SOUND_HRTF_R, WBUT_DISABLE);
-		widgSetTip(psWScreen, FRONTEND_SOUND_HRTF_R, _("HRTF is not supported on your device / system / OpenAL library"));
-	}
-
-	// map zoom
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MAP_ZOOM, _("Map Zoom"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_MAP_ZOOM_R, audioAndZoomOptionsMapZoomString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// map zoom rate
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MAP_ZOOM_RATE, _("Map Zoom Rate"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_MAP_ZOOM_RATE_R, audioAndZoomOptionsMapZoomRateString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// radar zoom
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_RADAR_ZOOM, _("Radar Zoom"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_RADAR_ZOOM_R, audioAndZoomOptionsRadarZoomString(), WBUT_SECONDARY)));
-	row.start++;
-
-	grid->setGeometry(0, 0, FRONTEND_BUTWIDTH, grid->idealHeight());
-
-	auto scrollableList = ScrollableListWidget::make();
-	scrollableList->setGeometry(0, FRONTEND_POS2Y, FRONTEND_BOTFORMW - 1, FRONTEND_BOTFORMH - FRONTEND_POS2Y - 1);
-	scrollableList->addItem(grid);
-	parent->attach(scrollableList);
-
-	// quit.
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-
-	//add some text down the side of the form
-	// TRANSLATORS: "AUDIO" options determine the volume of game sounds.
-	// "OPTIONS" means "SETTINGS".
-	// To break this message into two lines, you can use the delimiter "\n",
-	// e.g. "AUDIO / ZOOM\nOPTIONS" would show "OPTIONS" in a second line.
-	WzString messageString = WzString::fromUtf8(_("AUDIO / ZOOM OPTIONS"));
-	std::vector<WzString> messageStringLines = messageString.split("\n");
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, messageStringLines[0].toUtf8().c_str());
-	// show a second sidetext line if the translation requires it
-	if (messageStringLines.size() > 1)
-	{
-		messageString.remove(0, messageStringLines[0].length() + 1);
-		addSideText(FRONTEND_MULTILINE_SIDETEXT, FRONTEND_SIDEX + 22, \
-		FRONTEND_SIDEY, messageString.toUtf8().c_str());
-	}
-}
-
-bool runAudioAndZoomOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_FX:
-	case FRONTEND_3D_FX:
-	case FRONTEND_MUSIC:
-		break;
-
-	case FRONTEND_FX_SL:
-		sound_SetUIVolume((float)widgGetSliderPos(psWScreen, FRONTEND_FX_SL) / 100.0f);
-		break;
-
-	case FRONTEND_3D_FX_SL:
-		sound_SetEffectsVolume((float)widgGetSliderPos(psWScreen, FRONTEND_3D_FX_SL) / 100.0f);
-		break;
-
-	case FRONTEND_MUSIC_SL:
-		sound_SetMusicVolume((float)widgGetSliderPos(psWScreen, FRONTEND_MUSIC_SL) / 100.0f);
-		break;
-
-	case FRONTEND_SUBTITLES:
-	case FRONTEND_SUBTITLES_R:
-		seq_SetSubtitles(!seq_GetSubtitles());
-		widgSetString(psWScreen, FRONTEND_SUBTITLES_R, graphicsOptionsSubtitlesString());
-		break;
-
-	case FRONTEND_SOUND_HRTF:
-	case FRONTEND_SOUND_HRTF_R:
-		{
-			std::vector<HRTFMode> modesToCycle = { HRTFMode::Disabled, HRTFMode::Enabled, HRTFMode::Auto };
-			auto current = std::find(modesToCycle.begin(), modesToCycle.end(), war_GetHRTFMode());
-			if (current == modesToCycle.end())
-			{
-				current = modesToCycle.begin();
-			}
-			auto startingPoint = current;
-			bool successfulChange = false;
-			do {
-				current = seqCycle(current, modesToCycle.begin(), 1, modesToCycle.end() - 1);
-			} while ( (current != startingPoint) && ((successfulChange = sound_SetHRTFMode(*current)) == false) );
-			if (successfulChange)
-			{
-				war_SetHRTFMode(*current);
-				widgSetString(psWScreen, FRONTEND_SOUND_HRTF_R, audioAndZoomOptionsSoundHRTFMode().c_str());
-			}
-			break;
-		}
-
-	case FRONTEND_MAP_ZOOM:
-	case FRONTEND_MAP_ZOOM_R:
-		{
-		    war_SetMapZoom(seqCycle(war_GetMapZoom(), MINDISTANCE_CONFIG, MAP_ZOOM_CONFIG_STEP, MAXDISTANCE));
-		    widgSetString(psWScreen, FRONTEND_MAP_ZOOM_R, audioAndZoomOptionsMapZoomString().c_str());
-		    break;
-		}
-
-	case FRONTEND_MAP_ZOOM_RATE:
-	case FRONTEND_MAP_ZOOM_RATE_R:
-		{
-		    war_SetMapZoomRate(seqCycle(war_GetMapZoomRate(), MAP_ZOOM_RATE_MIN, MAP_ZOOM_RATE_STEP, MAP_ZOOM_RATE_MAX));
-		    widgSetString(psWScreen, FRONTEND_MAP_ZOOM_RATE_R, audioAndZoomOptionsMapZoomRateString().c_str());
-		    break;
-		}
-
-	case FRONTEND_RADAR_ZOOM:
-	case FRONTEND_RADAR_ZOOM_R:
-		{
-		    war_SetRadarZoom(seqCycle(war_GetRadarZoom(), MIN_RADARZOOM, RADARZOOM_STEP, MAX_RADARZOOM));
-		    widgSetString(psWScreen, FRONTEND_RADAR_ZOOM_R, audioAndZoomOptionsRadarZoomString().c_str());
-		    break;
-		}
-
-	case FRONTEND_QUIT:
-		changeTitleMode(OPTIONS);
-		break;
-
-	default:
-		break;
-	}
-
-	// If close button pressed then return from this menu.
-	if (CancelPressed())
-	{
-		changeTitleMode(OPTIONS);
-	}
-
-	widgDisplayScreen(psWScreen);						// show the widgets currently running
-
-	return true;
-}
-
-static char const *videoOptionsResolutionGetReadOnlyTooltip()
-{
-	switch (war_getWindowMode())
-	{
-		case WINDOW_MODE::desktop_fullscreen:
-			return _("In Desktop Fullscreen mode, the resolution matches that of your desktop \n(or what the window manager allows).");
-		case WINDOW_MODE::windowed:
-			return _("You can change the resolution by resizing the window normally. (Try dragging a corner / edge.)");
-		default:
-			return "";
-	}
-}
-
-static void videoOptionsDisableResolutionButtons()
-{
-	widgReveal(psWScreen, FRONTEND_RESOLUTION_READONLY_LABEL_CONTAINER);
-	widgReveal(psWScreen, FRONTEND_RESOLUTION_READONLY_CONTAINER);
-	auto readonlyResolutionTooltip = videoOptionsResolutionGetReadOnlyTooltip();
-	widgSetTip(psWScreen, FRONTEND_RESOLUTION_READONLY_LABEL, readonlyResolutionTooltip);
-	widgSetTip(psWScreen, FRONTEND_RESOLUTION_READONLY, readonlyResolutionTooltip);
-	widgHide(psWScreen, FRONTEND_RESOLUTION_DROPDOWN_LABEL_CONTAINER);
-	widgHide(psWScreen, FRONTEND_RESOLUTION_DROPDOWN);
-}
-
-static void videoOptionsEnableResolutionButtons()
-{
-	widgHide(psWScreen, FRONTEND_RESOLUTION_READONLY_LABEL_CONTAINER);
-	widgHide(psWScreen, FRONTEND_RESOLUTION_READONLY_CONTAINER);
-	widgReveal(psWScreen, FRONTEND_RESOLUTION_DROPDOWN_LABEL_CONTAINER);
-	widgReveal(psWScreen, FRONTEND_RESOLUTION_DROPDOWN);
-}
-
-static char const *videoOptionsWindowModeString()
-{
-	switch (war_getWindowMode())
-	{
-		case WINDOW_MODE::desktop_fullscreen:
-			return _("Desktop Full");
-		case WINDOW_MODE::windowed:
-			return _("Windowed");
-		case WINDOW_MODE::fullscreen:
-			return _("Fullscreen");
-	}
-	return "n/a"; // silence warning
-}
-
-static std::string videoOptionsAntialiasingString()
-{
-	if (war_getAntialiasing() == 0)
-	{
-		return _("Off");
-	}
-	else
-	{
-		return std::to_string(war_getAntialiasing()) + "×";
-	}
-}
-
-char const *videoOptionsWindowModeLabel()
-{
-	return _("Graphics Mode");
-}
-
-static char const *videoOptionsResolutionLabel()
-{
-	return _("Resolution");
-}
-
-char const *videoOptionsDisplayScaleLabel()
-{
-	return _("Display Scale");
-}
-
-static std::string videoOptionsTextureSizeString()
-{
-	char textureSize[20];
-	ssprintf(textureSize, "%d", getTextureSize());
-	return textureSize;
-}
-
-gfx_api::context::swap_interval_mode getCurrentSwapMode()
-{
-	return to_swap_mode(war_GetVsync());
-}
-
-void saveCurrentSwapMode(gfx_api::context::swap_interval_mode mode)
-{
-	war_SetVsync(to_int(mode));
-}
-
-char const *videoOptionsVsyncString()
-{
-	switch (getCurrentSwapMode()) {
-		case gfx_api::context::swap_interval_mode::immediate:
-			return _("Off");
-		case gfx_api::context::swap_interval_mode::vsync:
-			return _("On");
-		case gfx_api::context::swap_interval_mode::adaptive_vsync:
-			return _("Adaptive");
-	}
-	return "n/a";
-}
-
-std::string videoOptionsDisplayScaleString()
-{
-	char resolution[100];
-	ssprintf(resolution, "%d%%", war_GetDisplayScale());
-	return resolution;
-}
-
-std::string videoOptionsGfxBackendString()
-{
-	return to_display_string(war_getGfxBackend());
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Video Options
-
-class ScreenResolutionsModel
-{
-public:
-	typedef const std::vector<screeninfo>::iterator iterator;
-	typedef const std::vector<screeninfo>::const_iterator const_iterator;
-
-	ScreenResolutionsModel(): modes(ScreenResolutionsModel::loadModes())
-	{
-	}
-
-	const_iterator begin() const
-	{
-		return modes.begin();
-	}
-
-	const_iterator end() const
-	{
-		return modes.end();
-	}
-
-	const_iterator findResolutionClosestToCurrent() const
-	{
-		auto currentResolution = getCurrentResolution();
-		auto closest = std::lower_bound(modes.begin(), modes.end(), currentResolution, compareLess);
-		if (closest != modes.end() && compareEq(*closest, currentResolution))
-		{
-			return closest;
-		}
-
-		if (closest != modes.begin())
-		{
-			--closest;  // If current resolution doesn't exist, round down to next-highest one.
-		}
-
-		return closest;
-	}
-
-	void selectAt(size_t index) const
-	{
-		// Disable the ability to use the Video options menu to live-change the window size when in windowed mode.
-		// Why?
-		//	- Certain window managers don't report their own changes to window size through SDL in all circumstances.
-		//	  (For example, attempting to set a window size of 800x600 might result in no actual change when using a
-		//	   tiling window manager (ex. i3), but SDL thinks the window size has been set to 800x600. This obviously
-		//     breaks things.)
-		//  - Manual window resizing is supported (so there is no need for this functionality in the Video menu).
-		ASSERT_OR_RETURN(, wzGetCurrentWindowMode() == WINDOW_MODE::fullscreen, "Attempt to change resolution in windowed-mode / desktop-fullscreen mode");
-
-		ASSERT_OR_RETURN(, index < modes.size(), "Invalid mode index passed to ScreenResolutionsModel::selectAt");
-
-		auto selectedResolution = modes.at(index);
-
-		auto currentResolution = getCurrentResolution();
-		// Attempt to change the resolution
-		if (!wzChangeWindowResolution(selectedResolution.screen, selectedResolution.width, selectedResolution.height))
-		{
-			debug(
-				LOG_WARNING,
-				"Failed to change active resolution from: %s to: %s",
-				ScreenResolutionsModel::resolutionString(currentResolution).c_str(),
-				ScreenResolutionsModel::resolutionString(selectedResolution).c_str()
-			);
-		}
-
-		// Store the new width and height
-		war_SetScreen(selectedResolution.screen);
-		war_SetWidth(selectedResolution.width);
-		war_SetHeight(selectedResolution.height);
-
-		// Update the widget(s)
-		refreshCurrentVideoOptionsValues();
-	}
-
-	static std::string currentResolutionString()
-	{
-		return ScreenResolutionsModel::resolutionString(getCurrentResolution());
-	}
-
-	static std::string resolutionString(const screeninfo &info)
-	{
-		return astringf("[%d] %d × %d", info.screen, info.width, info.height);
-	}
-
-private:
-	const std::vector<screeninfo> modes;
-
-	static screeninfo getCurrentResolution()
-	{
-		screeninfo info;
-		info.screen = war_GetScreen();
-		info.width = war_GetWidth();
-		info.height = war_GetHeight();
-		info.refresh_rate = -1;  // Unused.
-		return info;
-	}
-
-	static std::vector<screeninfo> loadModes()
-	{
-		// Get resolutions, sorted with duplicates removed.
-		std::vector<screeninfo> modes = wzAvailableResolutions();
-		std::sort(modes.begin(), modes.end(), ScreenResolutionsModel::compareLess);
-		modes.erase(std::unique(modes.begin(), modes.end(), ScreenResolutionsModel::compareEq), modes.end());
-
-		return modes;
-	}
-
-	static std::tuple<int, int, int> compareKey(const screeninfo &x)
-	{
-		return std::make_tuple(x.screen, x.width, x.height);
-	}
-
-	static bool compareLess(const screeninfo &a, const screeninfo &b)
-	{
-		return ScreenResolutionsModel::compareKey(a) < ScreenResolutionsModel::compareKey(b);
-	}
-
-	static bool compareEq(const screeninfo &a, const screeninfo &b)
-	{
-		return ScreenResolutionsModel::compareKey(a) == ScreenResolutionsModel::compareKey(b);
-	}
-};
-
-void refreshCurrentVideoOptionsValues()
-{
-	widgSetString(psWScreen, FRONTEND_WINDOWMODE_R, videoOptionsWindowModeString());
-	widgSetString(psWScreen, FRONTEND_FSAA_R, videoOptionsAntialiasingString().c_str());
-	if (widgGetFromID(psWScreen, FRONTEND_RESOLUTION_READONLY)) // Resolution option may not be available
-	{
-		widgSetString(psWScreen, FRONTEND_RESOLUTION_READONLY, ScreenResolutionsModel::currentResolutionString().c_str());
-		if (war_getWindowMode() != WINDOW_MODE::fullscreen)
-		{
-			// If live window resizing is supported & the current mode is "windowed", disable the Resolution option and add a tooltip
-			// explaining the user can now resize the window normally.
-			videoOptionsDisableResolutionButtons();
-		}
-		else
-		{
-			videoOptionsEnableResolutionButtons();
-		}
-	}
-	widgSetString(psWScreen, FRONTEND_TEXTURESZ_R, videoOptionsTextureSizeString().c_str());
-	widgSetString(psWScreen, FRONTEND_VSYNC_R, videoOptionsVsyncString());
-	if (widgGetFromID(psWScreen, FRONTEND_DISPLAYSCALE_R)) // Display Scale option may not be available
-	{
-		widgSetString(psWScreen, FRONTEND_DISPLAYSCALE_R, videoOptionsDisplayScaleString().c_str());
-	}
-}
-
-static std::shared_ptr<WIDGET> makeResolutionDropdown()
-{
-	auto dropdown = std::make_shared<DropdownWidget>();
-	dropdown->id = FRONTEND_RESOLUTION_DROPDOWN;
-	dropdown->setListHeight(FRONTEND_BUTHEIGHT * 5);
-	const auto paddingSize = 10;
-
-	ScreenResolutionsModel screenResolutionsModel;
-	for (auto resolution: screenResolutionsModel)
-	{
-		auto item = makeTextButton(0, ScreenResolutionsModel::resolutionString(resolution), 0);
-		dropdown->addItem(Margin(0, paddingSize).wrap(item));
-	}
-
-	auto closestResolution = screenResolutionsModel.findResolutionClosestToCurrent();
-	if (closestResolution != screenResolutionsModel.end())
-	{
-		dropdown->setSelectedIndex(closestResolution - screenResolutionsModel.begin());
-	}
-
-	dropdown->setOnChange([screenResolutionsModel](DropdownWidget& dropdown) {
-		if (auto selectedIndex = dropdown.getSelectedIndex())
-		{
-			screenResolutionsModel.selectAt(selectedIndex.value());
-		}
-	});
-
-	return Margin(0, -paddingSize).wrap(dropdown);
-}
-
-void startVideoOptionsMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	// Add a note about changes taking effect on restart for certain options
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BOTFORM);
-
-	auto label = std::make_shared<W_LABEL>();
-	parent->attach(label);
-	label->setGeometry(FRONTEND_POS1X + 48, FRONTEND_POS1Y, FRONTEND_BUTWIDTH - FRONTEND_POS1X - 48, FRONTEND_BUTHEIGHT);
-	label->setFontColour(WZCOL_TEXT_BRIGHT);
-	label->setString(_("* Takes effect on game restart"));
-	label->setTextAlignment(WLAB_ALIGNBOTTOMLEFT);
-
-	auto grid = std::make_shared<GridLayout>();
-	grid_allocation::slot row(0);
-
-	// Fullscreen/windowed
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_WINDOWMODE, videoOptionsWindowModeLabel(), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_WINDOWMODE_R, videoOptionsWindowModeString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Resolution
-	auto resolutionReadonlyLabel = makeTextButton(FRONTEND_RESOLUTION_READONLY_LABEL, videoOptionsResolutionLabel(), WBUT_SECONDARY | WBUT_DISABLE);
-	resolutionReadonlyLabel->setTip(videoOptionsResolutionGetReadOnlyTooltip());
-	auto resolutionReadonlyLabelContainer = addMargin(resolutionReadonlyLabel);
-	resolutionReadonlyLabelContainer->id = FRONTEND_RESOLUTION_READONLY_LABEL_CONTAINER;
-
-	auto resolutionReadonlyValue = makeTextButton(FRONTEND_RESOLUTION_READONLY, ScreenResolutionsModel::currentResolutionString(), WBUT_SECONDARY | WBUT_DISABLE);
-	resolutionReadonlyValue->setTip(videoOptionsResolutionGetReadOnlyTooltip());
-	auto resolutionReadonlyValueContainer = addMargin(resolutionReadonlyValue);
-	resolutionReadonlyValueContainer->id = FRONTEND_RESOLUTION_READONLY_CONTAINER;
-
-	grid->place({0}, row, resolutionReadonlyLabelContainer);
-	grid->place({1, 1, false}, row, resolutionReadonlyValueContainer);
-
-	auto resolutionDropdownLabel = makeTextButton(FRONTEND_RESOLUTION_DROPDOWN_LABEL, videoOptionsResolutionLabel(), WBUT_SECONDARY);
-	auto resolutionDropdownLabelContainer = addMargin(resolutionDropdownLabel);
-	resolutionDropdownLabelContainer->id = FRONTEND_RESOLUTION_DROPDOWN_LABEL_CONTAINER;
-	grid->place({0}, row, resolutionDropdownLabelContainer);
-	grid->place({1, 1, false}, row, addMargin(makeResolutionDropdown()));
-	row.start++;
-
-	// Texture size
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_TEXTURESZ, _("Texture size"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_TEXTURESZ_R, videoOptionsTextureSizeString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Vsync
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_VSYNC, _("Vertical sync"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_VSYNC_R, videoOptionsVsyncString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Antialiasing
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_FSAA, _("Antialiasing*"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_FSAA_R, videoOptionsAntialiasingString(), WBUT_SECONDARY)));
-	row.start++;
-
-	auto antialiasing_label = std::make_shared<W_LABEL>();
-	parent->attach(antialiasing_label);
-	antialiasing_label->setGeometry(FRONTEND_POS1X + 48, FRONTEND_POS1Y - 18, FRONTEND_BUTWIDTH - FRONTEND_POS1X - 48, FRONTEND_BUTHEIGHT);
-	antialiasing_label->setFontColour(WZCOL_YELLOW);
-	antialiasing_label->setString(_("Warning: Antialiasing can cause crashes, especially with values > 16"));
-	antialiasing_label->setTextAlignment(WLAB_ALIGNBOTTOMLEFT);
-
-	// Display Scale
-	const bool showDisplayScale = wzAvailableDisplayScales().size() > 1;
-	if (showDisplayScale)
-	{
-		grid->place({0}, row, addMargin(makeTextButton(FRONTEND_DISPLAYSCALE, videoOptionsDisplayScaleLabel(), WBUT_SECONDARY)));
-		grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_DISPLAYSCALE_R, videoOptionsDisplayScaleString(), WBUT_SECONDARY)));
-		row.start++;
-	}
-
-	// Gfx Backend
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_GFXBACKEND, _("Graphics Backend*"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_GFXBACKEND_R, videoOptionsGfxBackendString(), WBUT_SECONDARY)));
-	row.start++;
-
-	grid->setGeometry(0, 0, FRONTEND_BUTWIDTH, grid->idealHeight());
-
-	auto scrollableList = ScrollableListWidget::make();
-	scrollableList->setGeometry(0, FRONTEND_POS2Y, FRONTEND_BOTFORMW - 1, FRONTEND_BOTFORMH - FRONTEND_POS2Y - 1);
-	scrollableList->addItem(grid);
-	parent->attach(scrollableList);
-
-	// Add some text down the side of the form
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("VIDEO OPTIONS"));
-
-	// Quit/return
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-
-	refreshCurrentVideoOptionsValues();
-}
-
-std::vector<unsigned int> availableDisplayScalesSorted()
-{
-	std::vector<unsigned int> displayScales = wzAvailableDisplayScales();
-	std::sort(displayScales.begin(), displayScales.end());
-	return displayScales;
-}
-
-void seqWindowMode()
-{
-	auto currentFullscreenMode = war_getWindowMode();
-	bool success = false;
-
-	auto supportedFullscreenModes = wzSupportedWindowModes();
-	if (supportedFullscreenModes.empty()) { return; }
-	size_t currentFullscreenModePos = 0;
-	for (size_t i = 0; i < supportedFullscreenModes.size(); i++)
-	{
-		if (supportedFullscreenModes[i] == currentFullscreenMode)
-		{
-			currentFullscreenModePos = i;
-			break;
-		}
-	}
-	auto startingFullscreenModePos = currentFullscreenModePos;
-
-	do
-	{
-		currentFullscreenModePos = seqCycle(currentFullscreenModePos, static_cast<size_t>(0), 1, supportedFullscreenModes.size() - 1);
-		success = wzChangeWindowMode(supportedFullscreenModes[currentFullscreenModePos]);
-
-	} while ((!success) && (currentFullscreenModePos != startingFullscreenModePos));
-
-	if (currentFullscreenModePos == startingFullscreenModePos)
-	{
-		// Failed to change window mode - display messagebox
-		wzDisplayDialog(Dialog_Warning, _("Unable to change Window Mode"), _("Warzone failed to change the Window mode.\nYour system / drivers may not support other modes."));
-	}
-	else if (success)
-	{
-		// succeeded changing vsync mode
-		war_setWindowMode(supportedFullscreenModes[currentFullscreenModePos]);
-	}
-}
-
-void seqVsyncMode()
-{
-	gfx_api::context::swap_interval_mode currentVsyncMode = getCurrentSwapMode();
-	auto startingVsyncMode = currentVsyncMode;
-	bool success = false;
-
-	do
-	{
-		currentVsyncMode = static_cast<gfx_api::context::swap_interval_mode>(seqCycle(static_cast<std::underlying_type<gfx_api::context::swap_interval_mode>::type>(currentVsyncMode), static_cast<std::underlying_type<gfx_api::context::swap_interval_mode>::type>(gfx_api::context::min_swap_interval_mode), 1, static_cast<std::underlying_type<gfx_api::context::swap_interval_mode>::type>(gfx_api::context::max_swap_interval_mode)));
-
-		success = gfx_api::context::get().setSwapInterval(currentVsyncMode);
-
-	} while ((!success) && (currentVsyncMode != startingVsyncMode));
-
-	if (currentVsyncMode == startingVsyncMode)
-	{
-		// Failed to change vsync mode - display messagebox
-		wzDisplayDialog(Dialog_Warning, _("Unable to change Vertical Sync"), _("Warzone failed to change the Vertical Sync mode.\nYour system / drivers may not support other modes."));
-	}
-	else if (success)
-	{
-		// succeeded changing vsync mode
-		saveCurrentSwapMode(currentVsyncMode);
-	}
-}
-
-void seqDisplayScale()
-{
-	std::vector<unsigned int> displayScales = availableDisplayScalesSorted();
-	assert(!displayScales.empty());
-
-	// Get currently-configured display scale.
-	unsigned int current_displayScale = war_GetDisplayScale();
-
-	// Find current display scale in list.
-	auto current = std::lower_bound(displayScales.begin(), displayScales.end(), current_displayScale);
-	if (current == displayScales.end() || *current != current_displayScale)
-	{
-		--current;  // If current display scale doesn't exist, round down to next-highest one.
-	}
-
-	// Increment/decrement and loop if required.
-	auto startingDisplayScale = current;
-	bool successfulDisplayScaleChange = false;
-	current = seqCycle(current, displayScales.begin(), 1, displayScales.end() - 1);
-
-	while (current != startingDisplayScale)
-	{
-		// Attempt to change the display scale
-		if (!wzChangeDisplayScale(*current))
-		{
-			debug(LOG_WARNING, "Failed to change display scale from: %d to: %d", current_displayScale, *current);
-
-			// try the next display scale factor, and loop
-			current = seqCycle(current, displayScales.begin(), 1, displayScales.end() - 1);
-			continue;
-		}
-		else
-		{
-			successfulDisplayScaleChange = true;
-			break;
-		}
-	}
-
-	if (!successfulDisplayScaleChange)
-		return;
-
-	// Store the new display scale
-	war_SetDisplayScale(*current);
-}
-
-bool runVideoOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_WINDOWMODE:
-	case FRONTEND_WINDOWMODE_R:
-		{
-			seqWindowMode();
-
-			// Update the widget(s)
-			refreshCurrentVideoOptionsValues();
-		}
-		break;
-
-	case FRONTEND_FSAA:
-	case FRONTEND_FSAA_R:
-		{
-			war_setAntialiasing(pow2Cycle(war_getAntialiasing(), 0, pie_GetMaxAntialiasing()));
-			widgSetString(psWScreen, FRONTEND_FSAA_R, videoOptionsAntialiasingString().c_str());
-			break;
-		}
-
-	case FRONTEND_TEXTURESZ:
-	case FRONTEND_TEXTURESZ_R:
-		{
-			int newTexSize = pow2Cycle(getTextureSize(), 128, 2048);
-
-			// Set the new size
-			setTextureSize(newTexSize);
-
-			// Update the widget
-			widgSetString(psWScreen, FRONTEND_TEXTURESZ_R, videoOptionsTextureSizeString().c_str());
-
-			break;
-		}
-
-	case FRONTEND_VSYNC:
-	case FRONTEND_VSYNC_R:
-		seqVsyncMode();
-
-		// Update the widget(s)
-		refreshCurrentVideoOptionsValues();
-		break;
-
-	case FRONTEND_GFXBACKEND:
-	case FRONTEND_GFXBACKEND_R:
-		{
-			const std::vector<video_backend> availableBackends = wzAvailableGfxBackends();
-			if (availableBackends.size() >= 1)
-			{
-				auto current = std::find(availableBackends.begin(), availableBackends.end(), war_getGfxBackend());
-				if (current == availableBackends.end())
-				{
-					current = availableBackends.begin();
-				}
-				current = seqCycle(current, availableBackends.begin(), 1, availableBackends.end() - 1);
-				war_setGfxBackend(*current);
-				widgSetString(psWScreen, FRONTEND_GFXBACKEND_R, videoOptionsGfxBackendString().c_str());
-			}
-			else
-			{
-				debug(LOG_ERROR, "There must be at least one valid backend");
-			}
-			break;
-		}
-
-	case FRONTEND_DISPLAYSCALE:
-	case FRONTEND_DISPLAYSCALE_R:
-		seqDisplayScale();
-
-		// Update the widget(s)
-		refreshCurrentVideoOptionsValues();
-		break;
-
-	case FRONTEND_QUIT:
-		changeTitleMode(OPTIONS);
-		break;
-
-	default:
-		break;
-	}
-
-	if (CancelPressed())
-	{
-		changeTitleMode(OPTIONS);
-	}
-
-	widgDisplayScreen(psWScreen);
-
-	return true;
-}
-
-char const *mouseOptionsMflipString()
-{
-	return getInvertMouseStatus() ? _("On") : _("Off");
-}
-
-char const *mouseOptionsTrapString()
-{
-	return war_GetTrapCursor() ? _("On") : _("Off");
-}
-
-char const *mouseOptionsMbuttonsString()
-{
-	return getRightClickOrders() ? _("On") : _("Off");
-}
-
-char const *mouseOptionsMmrotateString()
-{
-	return getMiddleClickRotate() ? _("Middle Mouse") : _("Right Mouse");
-}
-
-char const *mouseOptionsCursorModeString()
-{
-	return war_GetColouredCursor() ? _("On") : _("Off");
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Mouse Options
-void startMouseOptionsMenu()
-{
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BOTFORM);
-
-	auto grid = std::make_shared<GridLayout>();
-	grid_allocation::slot row(0);
-
-	////////////
-	// mouseflip
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MFLIP, _("Reverse Rotation"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_MFLIP_R, mouseOptionsMflipString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Cursor trapping
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_TRAP, _("Trap Cursor"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_TRAP_R, mouseOptionsTrapString(), WBUT_SECONDARY)));
-	row.start++;
-
-	////////////
-	// left-click orders
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MBUTTONS, _("Switch Mouse Buttons"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_MBUTTONS_R, mouseOptionsMbuttonsString(), WBUT_SECONDARY)));
-	row.start++;
-
-	////////////
-	// middle-click rotate
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_MMROTATE, _("Rotate Screen"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_MMROTATE_R, mouseOptionsMmrotateString(), WBUT_SECONDARY)));
-	row.start++;
-
-	// Hardware / software cursor toggle
-	grid->place({0}, row, addMargin(makeTextButton(FRONTEND_CURSORMODE, _("Colored Cursors"), WBUT_SECONDARY)));
-	grid->place({1, 1, false}, row, addMargin(makeTextButton(FRONTEND_CURSORMODE_R, mouseOptionsCursorModeString(), WBUT_SECONDARY)));
-	row.start++;
-
-	grid->setGeometry(0, 0, FRONTEND_BUTWIDTH, grid->idealHeight());
-
-	auto scrollableList = ScrollableListWidget::make();
-	scrollableList->setGeometry(0, FRONTEND_POS2Y, FRONTEND_BOTFORMW - 1, FRONTEND_BOTFORMH - FRONTEND_POS2Y - 1);
-	scrollableList->addItem(grid);
-	parent->attach(scrollableList);
-
-	// Add some text down the side of the form
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("MOUSE OPTIONS"));
-
-	// Quit/return
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-}
-
-bool runMouseOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_MFLIP:
-	case FRONTEND_MFLIP_R:
-		setInvertMouseStatus(!getInvertMouseStatus());
-		widgSetString(psWScreen, FRONTEND_MFLIP_R, mouseOptionsMflipString());
-		break;
-	case FRONTEND_TRAP:
-	case FRONTEND_TRAP_R:
-		war_SetTrapCursor(!war_GetTrapCursor());
-		widgSetString(psWScreen, FRONTEND_TRAP_R, mouseOptionsTrapString());
-		break;
-
-	case FRONTEND_MBUTTONS:
-	case FRONTEND_MBUTTONS_R:
-		setRightClickOrders(!getRightClickOrders());
-		widgSetString(psWScreen, FRONTEND_MBUTTONS_R, mouseOptionsMbuttonsString());
-		break;
-
-	case FRONTEND_MMROTATE:
-	case FRONTEND_MMROTATE_R:
-		setMiddleClickRotate(!getMiddleClickRotate());
-		widgSetString(psWScreen, FRONTEND_MMROTATE_R, mouseOptionsMmrotateString());
-		break;
-
-	case FRONTEND_CURSORMODE:
-	case FRONTEND_CURSORMODE_R:
-		war_SetColouredCursor(!war_GetColouredCursor());
-		widgSetString(psWScreen, FRONTEND_CURSORMODE_R, mouseOptionsCursorModeString());
-		wzSetCursor(CURSOR_DEFAULT);
-		break;
-
-	case FRONTEND_QUIT:
-		changeTitleMode(OPTIONS);
-		break;
-
-	default:
-		break;
-	}
-
-	if (CancelPressed())
-	{
-		changeTitleMode(OPTIONS);
-	}
-
-	widgDisplayScreen(psWScreen);
-
-	return true;
-}
-
-static char const *gameOptionsDifficultyString()
-{
-	switch (getDifficultyLevel())
-	{
-	case DL_EASY: return _("Easy");
-	case DL_NORMAL: return _("Normal");
-	case DL_HARD: return _("Hard");
-	case DL_INSANE: return _("Insane");
-	}
-	return _("Unsupported");
-}
-
-static std::string gameOptionsCameraSpeedString()
-{
-	char cameraSpeed[20];
-	if(getCameraAccel())
-	{
-		ssprintf(cameraSpeed, "%d", war_GetCameraSpeed());
-	}
-	else
-	{
-		ssprintf(cameraSpeed, "%d / 2", war_GetCameraSpeed());
-	}
-	return cameraSpeed;
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Game Options Menu
-void startGameOptionsMenu()
-{
-	UDWORD	w, h;
-	int playercolor;
-
-	addBackdrop();
-	addTopForm(false);
-	addBottomForm();
-
-	// language
-	addTextButton(FRONTEND_LANGUAGE,  FRONTEND_POS2X - 35, FRONTEND_POS2Y, _("Language"), WBUT_SECONDARY);
-	addTextButton(FRONTEND_LANGUAGE_R,  FRONTEND_POS2M - 55, FRONTEND_POS2Y, getLanguageName(), WBUT_SECONDARY);
-
-	// Difficulty
-	addTextButton(FRONTEND_DIFFICULTY,   FRONTEND_POS3X - 35, FRONTEND_POS3Y, _("Campaign Difficulty"), WBUT_SECONDARY);
-	addTextButton(FRONTEND_DIFFICULTY_R, FRONTEND_POS3M - 55, FRONTEND_POS3Y, gameOptionsDifficultyString(), WBUT_SECONDARY);
-
-	// Camera speed
-	addTextButton(FRONTEND_CAMERASPEED, FRONTEND_POS4X - 35, FRONTEND_POS4Y, _("Camera Speed"), WBUT_SECONDARY);
-	addTextButton(FRONTEND_CAMERASPEED_R, FRONTEND_POS4M - 55, FRONTEND_POS4Y, gameOptionsCameraSpeedString(), WBUT_SECONDARY);
-
-	// Colour stuff
-	addTextButton(FRONTEND_COLOUR, FRONTEND_POS5X - 35, FRONTEND_POS5Y, _("Unit Colour:"), 0);
-
-	w = iV_GetImageWidth(FrontImages, IMAGE_PLAYERN);
-	h = iV_GetImageHeight(FrontImages, IMAGE_PLAYERN);
-
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_P0, FRONTEND_POS6M + (0 * (w + 6)) -20, FRONTEND_POS6Y, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, 0);
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_P4, FRONTEND_POS6M + (1 * (w + 6)) -20, FRONTEND_POS6Y, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, 4);
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_P5, FRONTEND_POS6M + (2 * (w + 6)) -20, FRONTEND_POS6Y, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, 5);
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_P6, FRONTEND_POS6M + (3 * (w + 6)) -20, FRONTEND_POS6Y, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, 6);
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_P7, FRONTEND_POS6M + (4 * (w + 6)) -20, FRONTEND_POS6Y, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, 7);
-
-	// FIXME: if playercolor = 1-3, then we Assert in widgSetButtonState() since we don't define FE_P1 - FE_P3
-	// I assume the reason is that in SP games, those are reserved for the AI?  Valid values are 0, 4-7.
-	// This is a workaround, until we find what is setting that to 1-3.  See configuration.c:701
-	playercolor = war_GetSPcolor();
-	if (playercolor >= 1 && playercolor <= 3)
-	{
-		playercolor = 0;
-	}
-	widgSetButtonState(psWScreen, FE_P0 + playercolor, WBUT_LOCK);
-	addTextButton(FRONTEND_COLOUR_CAM, FRONTEND_POS6X - 20, FRONTEND_POS6Y, _("Campaign"), 0);
-
-	playercolor = war_getMPcolour();
-	for (int colour = -1; colour < MAX_PLAYERS_IN_GUI; ++colour)
-	{
-		int cellX = (colour + 1) % 7;
-		int cellY = (colour + 1) / 7;
-		addMultiBut(psWScreen, FRONTEND_BOTFORM, FE_MP_PR + colour + 1, FRONTEND_POS7M + cellX * (w + 2) -20, FRONTEND_POS7Y + cellY * (h + 2) - 5, w, h, nullptr, IMAGE_PLAYERN, IMAGE_PLAYERX, true, colour >= 0 ? colour : MAX_PLAYERS + 1);
-	}
-	widgSetButtonState(psWScreen, FE_MP_PR + playercolor + 1, WBUT_LOCK);
-	addTextButton(FRONTEND_COLOUR_MP, FRONTEND_POS7X - 20, FRONTEND_POS7Y, _("Skirmish/Multiplayer"), 0);
-
-	// Quit
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, FRONTEND_QUIT, 10, 10, 30, 29, P_("menu", "Return"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);
-
-	// "Help Us Translate" link
-	const auto helpTranslateMessage = astringf(_("Help us improve translations of Warzone 2100: %s"), TRANSLATION_URL);
-	addSmallTextButton(FRONTEND_HYPERLINK, FRONTEND_POS9X, FRONTEND_POS9Y, helpTranslateMessage.c_str(), 0);
-	widgSetTip(psWScreen, FRONTEND_HYPERLINK, _("Click to open webpage."));
-
-	// Add some text down the side of the form
-	addSideText(FRONTEND_SIDETEXT, FRONTEND_SIDEX, FRONTEND_SIDEY, _("GAME OPTIONS"));
-}
-
-bool runGameOptionsMenu()
-{
-	WidgetTriggers const &triggers = widgRunScreen(psWScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case FRONTEND_HYPERLINK:
-		openURLInBrowser(TRANSLATION_URL);
-		break;
-	case FRONTEND_LANGUAGE:
-	case FRONTEND_LANGUAGE_R:
-
-		setNextLanguage(widgGetButtonKey_DEPRECATED(psWScreen) == WKEY_PRIMARY);
-		widgSetString(psWScreen, FRONTEND_LANGUAGE_R, getLanguageName());
-		/* Hack to reset current menu text, which looks fancy. */
-		widgSetString(psWScreen, FRONTEND_SIDETEXT, _("GAME OPTIONS"));
-		widgGetFromID(psWScreen, FRONTEND_QUIT)->setTip(P_("menu", "Return"));
-		widgSetString(psWScreen, FRONTEND_LANGUAGE, _("Language"));
-		widgSetString(psWScreen, FRONTEND_COLOUR, _("Unit Colour:"));
-		widgSetString(psWScreen, FRONTEND_COLOUR_CAM, _("Campaign"));
-		widgSetString(psWScreen, FRONTEND_COLOUR_MP, _("Skirmish/Multiplayer"));
-		widgSetString(psWScreen, FRONTEND_DIFFICULTY, _("Campaign Difficulty"));
-		widgSetString(psWScreen, FRONTEND_CAMERASPEED, _("Camera Speed"));
-		widgSetString(psWScreen, FRONTEND_DIFFICULTY_R, gameOptionsDifficultyString());
-
-		// hack to update translations of AI names and tooltips
-		readAIs();
-		break;
-
-	case FRONTEND_DIFFICULTY:
-	case FRONTEND_DIFFICULTY_R:
-		setDifficultyLevel(seqCycle(getDifficultyLevel(), DL_EASY, 1, DL_INSANE));
-		widgSetString(psWScreen, FRONTEND_DIFFICULTY_R, gameOptionsDifficultyString());
-		if (getDifficultyLevel() == DL_INSANE)
-		{
-			const std::string DIFF_TAG = "difficulty";
-
-			if (!hasNotificationsWithTag(DIFF_TAG))
-			{
-				WZ_Notification notification;
-				notification.duration = 10 * GAME_TICKS_PER_SEC;;
-				notification.contentTitle = _("Insane Difficulty");
-				notification.contentText = _("This difficulty is for very experienced players!");
-				notification.tag = DIFF_TAG;
-				notification.largeIcon = WZ_Notification_Image("images/notifications/skull_crossbones.png");
-
-				addNotification(notification, WZ_Notification_Trigger::Immediate());
-			}
-		}
-		break;
-
-	case FRONTEND_CAMERASPEED:
-	case FRONTEND_CAMERASPEED_R:
-		war_SetCameraSpeed(seqCycle(war_GetCameraSpeed(), CAMERASPEED_MIN, CAMERASPEED_STEP, CAMERASPEED_MAX));
-		widgSetString(psWScreen, FRONTEND_CAMERASPEED_R, gameOptionsCameraSpeedString().c_str());
-		break;
-
-	case FRONTEND_QUIT:
-		changeTitleMode(OPTIONS);
-		break;
-
-	case FE_P0:
-		widgSetButtonState(psWScreen, FE_P0, WBUT_LOCK);
-		widgSetButtonState(psWScreen, FE_P4, 0);
-		widgSetButtonState(psWScreen, FE_P5, 0);
-		widgSetButtonState(psWScreen, FE_P6, 0);
-		widgSetButtonState(psWScreen, FE_P7, 0);
-		war_SetSPcolor(0);
-		break;
-	case FE_P4:
-		widgSetButtonState(psWScreen, FE_P0, 0);
-		widgSetButtonState(psWScreen, FE_P4, WBUT_LOCK);
-		widgSetButtonState(psWScreen, FE_P5, 0);
-		widgSetButtonState(psWScreen, FE_P6, 0);
-		widgSetButtonState(psWScreen, FE_P7, 0);
-		war_SetSPcolor(4);
-		break;
-	case FE_P5:
-		widgSetButtonState(psWScreen, FE_P0, 0);
-		widgSetButtonState(psWScreen, FE_P4, 0);
-		widgSetButtonState(psWScreen, FE_P5, WBUT_LOCK);
-		widgSetButtonState(psWScreen, FE_P6, 0);
-		widgSetButtonState(psWScreen, FE_P7, 0);
-		war_SetSPcolor(5);
-		break;
-	case FE_P6:
-		widgSetButtonState(psWScreen, FE_P0, 0);
-		widgSetButtonState(psWScreen, FE_P4, 0);
-		widgSetButtonState(psWScreen, FE_P5, 0);
-		widgSetButtonState(psWScreen, FE_P6, WBUT_LOCK);
-		widgSetButtonState(psWScreen, FE_P7, 0);
-		war_SetSPcolor(6);
-		break;
-	case FE_P7:
-		widgSetButtonState(psWScreen, FE_P0, 0);
-		widgSetButtonState(psWScreen, FE_P4, 0);
-		widgSetButtonState(psWScreen, FE_P5, 0);
-		widgSetButtonState(psWScreen, FE_P6, 0);
-		widgSetButtonState(psWScreen, FE_P7, WBUT_LOCK);
-		war_SetSPcolor(7);
-		break;
-
-	default:
-		break;
-	}
-
-	if (id >= FE_MP_PR && id <= FE_MP_PMAX)
-	{
-		int chosenColour = id - FE_MP_PR - 1;
-		for (int colour = -1; colour < MAX_PLAYERS_IN_GUI; ++colour)
-		{
-			unsigned thisID = FE_MP_PR + colour + 1;
-			widgSetButtonState(psWScreen, thisID, id == thisID ? WBUT_LOCK : 0);
-		}
-		war_setMPcolour(chosenColour);
-	}
-
-	// If close button pressed then return from this menu.
-	if (CancelPressed())
-	{
-		changeTitleMode(OPTIONS);
-	}
-
-	widgDisplayScreen(psWScreen);						// show the widgets currently running
-
-	return true;
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// ////////////////////////////////////////////////////////////////////////////
-// ////////////////////////////////////////////////////////////////////////////
-// common widgets.
-
-W_FORM *addBackdrop()
-{
-	return addBackdrop(psWScreen);
-}
-
-W_FORM *addBackdrop(const std::shared_ptr<W_SCREEN> &screen)
-{
-	ASSERT_OR_RETURN(nullptr, screen != nullptr, "Invalid screen pointer");
-	W_FORMINIT sFormInit;                              // Backdrop
-	sFormInit.formID = 0;
-	sFormInit.id = FRONTEND_BACKDROP;
-	sFormInit.style = WFORM_PLAIN;
-	sFormInit.calcLayout = LAMBDA_CALCLAYOUT_SIMPLE({
-		psWidget->setGeometry((SWORD)((pie_GetVideoBufferWidth() - HIDDEN_FRONTEND_WIDTH) / 2),
-							  (SWORD)((pie_GetVideoBufferHeight() - HIDDEN_FRONTEND_HEIGHT) / 2),
-							  HIDDEN_FRONTEND_WIDTH - 1,
-							  HIDDEN_FRONTEND_HEIGHT - 1);
-	});
-	sFormInit.pDisplay = displayTitleBitmap;
-	sFormInit.pUserData = new TitleBitmapCache();
-	sFormInit.onDelete = [](WIDGET *psWidget) {
-		assert(psWidget->pUserData != nullptr);
-		delete ((TitleBitmapCache *)psWidget->pUserData);
-		psWidget->pUserData = nullptr;
-	};
-
-	return widgAddForm(screen, &sFormInit);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-void addTopForm(bool wide)
-{
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BACKDROP);
-
-	auto topForm = std::make_shared<IntFormTransparent>();
-	parent->attach(topForm);
-	topForm->id = FRONTEND_TOPFORM;
-	if (wide)
-	{
-		topForm->setCalcLayout(LAMBDA_CALCLAYOUT_SIMPLE({
-			psWidget->setGeometry(FRONTEND_TOPFORM_WIDEX, FRONTEND_TOPFORM_WIDEY, FRONTEND_TOPFORM_WIDEW, FRONTEND_TOPFORM_WIDEH);
-		}));
-	}
-	else
-	{
-		topForm->setCalcLayout(LAMBDA_CALCLAYOUT_SIMPLE({
-			psWidget->setGeometry(FRONTEND_TOPFORMX, FRONTEND_TOPFORMY, FRONTEND_TOPFORMW, FRONTEND_TOPFORMH);
-		}));
-	}
-
-	W_FORMINIT sFormInit;
-	sFormInit.formID = FRONTEND_TOPFORM;
-	sFormInit.id	= FRONTEND_LOGO;
-	int imgW = iV_GetImageWidth(FrontImages, IMAGE_FE_LOGO);
-	int imgH = iV_GetImageHeight(FrontImages, IMAGE_FE_LOGO);
-	int dstW = topForm->width();
-	int dstH = topForm->height();
-	if (imgW * dstH < imgH * dstW) // Want to set aspect ratio dstW/dstH = imgW/imgH.
-	{
-		dstW = imgW * dstH / imgH; // Too wide.
-	}
-	else if (imgW * dstH > imgH * dstW)
-	{
-		dstH = imgH * dstW / imgW; // Too high.
-	}
-	sFormInit.x = (topForm->width()  - dstW) / 2;
-	sFormInit.y = (topForm->height() - dstH) / 2;
-	sFormInit.width  = dstW;
-	sFormInit.height = dstH;
-	sFormInit.pDisplay = displayLogo;
-	widgAddForm(psWScreen, &sFormInit);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-void addBottomForm()
-{
-	WIDGET *parent = widgGetFromID(psWScreen, FRONTEND_BACKDROP);
-
-	auto botForm = std::make_shared<IntFormAnimated>();
-	parent->attach(botForm);
-	botForm->id = FRONTEND_BOTFORM;
-	botForm->setCalcLayout(LAMBDA_CALCLAYOUT_SIMPLE({
-		psWidget->setGeometry(FRONTEND_BOTFORMX, FRONTEND_BOTFORMY, FRONTEND_BOTFORMW, FRONTEND_BOTFORMH);
-	}));
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-void addText(UDWORD id, UDWORD PosX, UDWORD PosY, const char *txt, UDWORD formID)
-{
-	WIDGET *parent = widgGetFromID(psWScreen, formID);
-
-	auto label = std::make_shared<W_LABEL>();
-	parent->attach(label);
-	label->id = id;
-	label->setGeometry(PosX, PosY, MULTIOP_READY_WIDTH, FRONTEND_BUTHEIGHT);
-	label->setTextAlignment(WLAB_ALIGNCENTRE);
-	label->setFont(font_small, WZCOL_TEXT_BRIGHT);
-	label->setString(txt);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-W_LABEL *addSideText(const std::shared_ptr<W_SCREEN> &psScreen, UDWORD formId, UDWORD id, UDWORD PosX, UDWORD PosY, const char *txt)
-{
-	ASSERT_OR_RETURN(nullptr, psScreen != nullptr, "Invalid screen pointer");
-
-	W_LABINIT sLabInit;
-
-	sLabInit.formID = formId;
-	sLabInit.id = id;
-	sLabInit.x = (short) PosX;
-	sLabInit.y = (short) PosY;
-	sLabInit.width = 30;
-	sLabInit.height = FRONTEND_BOTFORMH;
-
-	sLabInit.FontID = font_large;
-
-	sLabInit.pDisplay = displayTextAt270;
-	sLabInit.pText = WzString::fromUtf8(txt);
-	sLabInit.pUserData = new DisplayTextOptionCache();
-	sLabInit.onDelete = [](WIDGET *psWidget) {
-		assert(psWidget->pUserData != nullptr);
-		delete static_cast<DisplayTextOptionCache *>(psWidget->pUserData);
-		psWidget->pUserData = nullptr;
-	};
-
-	return widgAddLabel(psScreen, &sLabInit);
-}
-
-W_LABEL *addSideText(UDWORD id, UDWORD PosX, UDWORD PosY, const char *txt)
-{
-	return addSideText(psWScreen, FRONTEND_BACKDROP, id, PosX, PosY, txt);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-std::shared_ptr<W_BUTTON> makeTextButton(UDWORD id, const std::string &txt, unsigned int style)
-{
-	W_BUTINIT sButInit;
-
-	sButInit.formID = FRONTEND_BOTFORM;
-	sButInit.id = id;
-
-	// Align
-	if (!(style & WBUT_TXTCENTRE))
-	{
-		sButInit.width = (short)iV_GetTextWidth(txt.c_str(), font_large);
-	}
-	else
-	{
-		sButInit.style |= WBUT_TXTCENTRE;
-	}
-
-	// Enable right clicks
-	if (style & WBUT_SECONDARY)
-	{
-		sButInit.style |= WBUT_SECONDARY;
-	}
-
-	sButInit.UserData = (style & WBUT_DISABLE); // store disable state
-	sButInit.pUserData = new DisplayTextOptionCache();
-	sButInit.onDelete = [](WIDGET *psWidget) {
-		assert(psWidget->pUserData != nullptr);
-		delete static_cast<DisplayTextOptionCache *>(psWidget->pUserData);
-		psWidget->pUserData = nullptr;
-	};
-
-	sButInit.height = FRONTEND_BUTHEIGHT;
-	sButInit.pDisplay = displayTextOption;
-	sButInit.FontID = font_large;
-	sButInit.pText = txt.c_str();
-
-	auto button = std::make_shared<W_BUTTON>(&sButInit);
-	// Disable button
-	if (style & WBUT_DISABLE)
-	{
-		button->setState(WBUT_DISABLE);
-	}
-
-	return button;
-}
-
 
 
 // ////////////////////////////////////////////////////////////////////////////
