+++ src/ingameop.cpp
@@ -51,6 +51,9 @@
 #include "seqdisp.h"
 
 #include "uiscreens/common.h"
+#include "uiscreens/options/graphicsoptions.h"
+#include "uiscreens/options/videooptions.h"
+#include "uiscreens/options/mouseoptions.h"
 
 bool hostQuitConfirmation = true;
