+++ src/multimenu.cpp
@@ -66,6 +66,8 @@
 #include "loop.h"
 #include "frontend.h"
 
+#include "uiscreens/common.h"
+
 // ////////////////////////////////////////////////////////////////////////////
 // defines
