+++ src/uiscreens/options/autiozoomoptions.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../common.h"
+
+void startAudioAndZoomOptionsMenu();
+bool runAudioAndZoomOptionsMenu();
