+++ src/uiscreens/options/gameoptions.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../common.h"
+
+bool runGameOptionsMenu();
+void startGameOptionsMenu();
+
