+++ src/uiscreens/options/optionsmenu.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../common.h"
+
+void startOptionsMenu();
+bool runOptionsMenu();
