+++ src/uiscreens/titlemenu.h
@@ -2,7 +2,6 @@
 
 #include "common.h"
 
-
 void startTitleMenu();
 bool runTitleMenu();
 void runHyperlink();
